reorganize /data
