+++ data/dataSources/CharacterStatDefinitions.json
