+++ data/dataSources/EquipmentSlotDefinitions.json
