+++ data/dataSources/LoadoutSlots.json
