+++ data/dataSources/ProfileTypes.json
@@ -1,0 +1,162 @@
+[
+    {
+        "ID": 1,
+        "ICON_ID": 0,
+        "NAME_ID": 0
+    },
+    {
+        "ID": 2,
+        "ICON_ID": 0,
+        "NAME_ID": 0
+    },
+    {
+        "ID": 3,
+        "ICON_ID": 0,
+        "NAME_ID": 1097
+    },
+    {
+        "ID": 4,
+        "ICON_ID": 0,
+        "NAME_ID": 0
+    },
+    {
+        "ID": 5,
+        "ICON_ID": 0,
+        "NAME_ID": 0
+    },
+    {
+        "ID": 6,
+        "ICON_ID": 0,
+        "NAME_ID": 0
+    },
+    {
+        "ID": 7,
+        "ICON_ID": 0,
+        "NAME_ID": 0
+    },
+    {
+        "ID": 8,
+        "ICON_ID": 0,
+        "NAME_ID": 6000056
+    },
+    {
+        "ID": 9,
+        "ICON_ID": 0,
+        "NAME_ID": 63
+    },
+    {
+        "ID": 10,
+        "ICON_ID": 0,
+        "NAME_ID": 64
+    },
+    {
+        "ID": 11,
+        "ICON_ID": 0,
+        "NAME_ID": 60
+    },
+    {
+        "ID": 12,
+        "ICON_ID": 0,
+        "NAME_ID": 79
+    },
+    {
+        "ID": 13,
+        "ICON_ID": 0,
+        "NAME_ID": 1088
+    },
+    {
+        "ID": 14,
+        "ICON_ID": 0,
+        "NAME_ID": 696
+    },
+    {
+        "ID": 15,
+        "ICON_ID": 0,
+        "NAME_ID": 697
+    },
+    {
+        "ID": 16,
+        "ICON_ID": 0,
+        "NAME_ID": 1089
+    },
+    {
+        "ID": 17,
+        "ICON_ID": 0,
+        "NAME_ID": 1101
+    },
+    {
+        "ID": 18,
+        "ICON_ID": 0,
+        "NAME_ID": 1175
+    },
+    {
+        "ID": 19,
+        "ICON_ID": 0,
+        "NAME_ID": 1200
+    },
+    {
+        "ID": 20,
+        "ICON_ID": 0,
+        "NAME_ID": 1224
+    },
+    {
+        "ID": 21,
+        "ICON_ID": 0,
+        "NAME_ID": 1235
+    },
+    {
+        "ID": 22,
+        "ICON_ID": 0,
+        "NAME_ID": 1245
+    },
+    {
+        "ID": 23,
+        "ICON_ID": 0,
+        "NAME_ID": 1213
+    },
+    {
+        "ID": 24,
+        "ICON_ID": 0,
+        "NAME_ID": 1246
+    },
+    {
+        "ID": 25,
+        "ICON_ID": 0,
+        "NAME_ID": 1247
+    },
+    {
+        "ID": 26,
+        "ICON_ID": 0,
+        "NAME_ID": 21
+    },
+    {
+        "ID": 27,
+        "ICON_ID": 0,
+        "NAME_ID": 1391
+    },
+    {
+        "ID": 28,
+        "ICON_ID": 0,
+        "NAME_ID": 13
+    },
+    {
+        "ID": 29,
+        "ICON_ID": 0,
+        "NAME_ID": 8889
+    },
+    {
+        "ID": 30,
+        "ICON_ID": 0,
+        "NAME_ID": 8951
+    },
+    {
+        "ID": 31,
+        "ICON_ID": 0,
+        "NAME_ID": 0
+    },
+    {
+        "ID": 32,
+        "ICON_ID": 0,
+        "NAME_ID": 9014
+    }
+]
