+++ src/servers/LoginServer/loginserver.ts
@@ -110,7 +110,7 @@ export class LoginServer extends EventEmitter {
             case "CharacterSelectInfoRequest":
               let CharactersInfo;
               if (this._soloMode) {
-                const SinglePlayerCharacter = require("../../../data/single_player_character.json");
+                const SinglePlayerCharacter = require("../../../data/sampleData/single_player_character.json");
 
                 const cowboy = _.cloneDeep(SinglePlayerCharacter); // for fun 🤠
                 cowboy.characterId = getCharacterId(99);
@@ -146,7 +146,7 @@ export class LoginServer extends EventEmitter {
                 servers = await this._db.collection("servers").find().toArray();
               } else {
                 if (this._soloMode) {
-                  const SoloServer = require("../../../data/single_player_server.json");
+                  const SoloServer = require("../../../data/sampleData/single_player_server.json");
                   servers = [SoloServer];
                 }
               }
